azure-ai-images-classification/requirements.txt
python-dotenv
azure-cognitiveservices-vision-customvision==3.1.0